af-xdp: address compile warnings
runmode-af-xdp.c:119:52: error: implicit conversion loses integer precision: 'const int' to 'uint16_t' (aka 'unsigned short') [-Werror,-Wimplicit-int-conversion]
  119 |         aconf->threads = (nr_cores <= nr_queues) ? nr_cores : nr_queues;
      |                        ~                           ^~~~~~~~
runmode-af-xdp.c:119:63: error: implicit conversion loses integer precision: 'const int' to 'uint16_t' (aka 'unsigned short') [-Werror,-Wimplicit-int-conversion]
  119 |         aconf->threads = (nr_cores <= nr_queues) ? nr_cores : nr_queues;
      |                        ~                                      ^~~~~~~~~
runmode-af-xdp.c:135:26: error: implicit conversion loses integer precision: 'const int' to 'uint16_t' (aka 'unsigned short') [-Werror,-Wimplicit-int-conversion]
  135 |         aconf->threads = nr_queues;
      |                        ~ ^~~~~~~~~
runmode-af-xdp.c:268:41: error: implicit conversion loses integer precision: 'intmax_t' (aka 'long') to 'uint32_t' (aka 'unsigned int') [-Werror,-Wshorten-64-to-32]
  268 |                 aconf->busy_poll_time = conf_val_int;
      |                                       ~ ^~~~~~~~~~~~
runmode-af-xdp.c:275:43: error: implicit conversion loses integer precision: 'intmax_t' (aka 'long') to 'uint32_t' (aka 'unsigned int') [-Werror,-Wshorten-64-to-32]
  275 |                 aconf->busy_poll_budget = conf_val_int;
      |                                         ~ ^~~~~~~~~~~~
runmode-af-xdp.c:282:40: error: implicit conversion loses integer precision: 'intmax_t' (aka 'long') to 'uint32_t' (aka 'unsigned int') [-Werror,-Wshorten-64-to-32]
  282 |             aconf->gro_flush_timeout = conf_val_int;
      |                                      ~ ^~~~~~~~~~~~
runmode-af-xdp.c:287:43: error: implicit conversion loses integer precision: 'intmax_t' (aka 'long') to 'uint32_t' (aka 'unsigned int') [-Werror,-Wshorten-64-to-32]
  287 |             aconf->napi_defer_hard_irqs = conf_val_int;
      |                                         ~ ^~~~~~~~~~~~
7 errors generated.

source-af-xdp.c:426:24: error: implicit conversion loses integer precision: 'int' to 'uint8_t' (aka 'unsigned char') [-Werror,-Wimplicit-int-conversion]
  426 |     ptv->afxdp_state = state;
      |                      ~ ^~~~~
source-af-xdp.c:642:44: error: implicit conversion loses integer precision: 'uint32_t' (aka 'unsigned int') to '__u16' (aka 'unsigned short') [-Werror,-Wimplicit-int-conversion]
  642 |     ptv->xsk.cfg.bind_flags = afxdpconfig->bind_flags;
      |                             ~ ~~~~~~~~~~~~~^~~~~~~~~~
2 errors generated.

+++ src/runmode-af-xdp.c
@@ -109,11 +109,11 @@ static TmEcode ConfigSetThreads(AFXDPIfaceConfig *aconf, const char *entry_str)
         SCReturnInt(TM_ECODE_FAILED);
     }
 
-    const int nr_queues = GetIfaceRSSQueuesNum(aconf->iface);
+    const uint16_t nr_queues = (uint16_t)GetIfaceRSSQueuesNum(aconf->iface);
 
     if (strcmp(entry_str, "auto") == 0) {
 
-        const int nr_cores = (int)UtilCpuGetNumProcessorsOnline();
+        const uint16_t nr_cores = UtilCpuGetNumProcessorsOnline();
 
         /* Threads limited to MIN(cores vs queues) */
         aconf->threads = (nr_cores <= nr_queues) ? nr_cores : nr_queues;
@@ -131,7 +131,7 @@ static TmEcode ConfigSetThreads(AFXDPIfaceConfig *aconf, const char *entry_str)
 
     if (aconf->threads > nr_queues) {
         SCLogWarning(
-                "Selected threads greater than configured queues, using: %d thread(s)", nr_queues);
+                "Selected threads greater than configured queues, using: %u thread(s)", nr_queues);
         aconf->threads = nr_queues;
     }
 
@@ -265,26 +265,26 @@ static void *ParseAFXDPConfig(const char *iface)
         if (SCConfGetChildValueIntWithDefault(
                     if_root, if_default, "busy-poll-time", &conf_val_int) == 1) {
             if (conf_val_int) {
-                aconf->busy_poll_time = conf_val_int;
+                aconf->busy_poll_time = (uint32_t)conf_val_int;
             }
         }
 
         if (SCConfGetChildValueIntWithDefault(
                     if_root, if_default, "busy-poll-budget", &conf_val_int) == 1) {
             if (conf_val_int) {
-                aconf->busy_poll_budget = conf_val_int;
+                aconf->busy_poll_budget = (uint32_t)conf_val_int;
             }
         }
 
         /* 0 value is valid for these Linux tunable's */
         if (SCConfGetChildValueIntWithDefault(
                     if_root, if_default, "gro-flush-timeout", &conf_val_int) == 1) {
-            aconf->gro_flush_timeout = conf_val_int;
+            aconf->gro_flush_timeout = (uint32_t)conf_val_int;
         }
 
         if (SCConfGetChildValueIntWithDefault(
                     if_root, if_default, "napi-defer-hard-irq", &conf_val_int) == 1) {
-            aconf->napi_defer_hard_irqs = conf_val_int;
+            aconf->napi_defer_hard_irqs = (uint32_t)conf_val_int;
         }
     }
 #endif
